Fixes idle_timeout property string and disables the default.
This is essentially an aggregation of the changes from
https://github.com/pepperdata/asynchbase/compare/master_pepperdata...pepperdata:bpitman?expand=1
 - Commit 7b6c492ee625414a1ac52722957021ce183eb12a
 - Commit c13bcd2827375562cf35c72ddcb5fadc87aecaaa
 - Commit 508a2cbf604856e0355ec54ad9e394774fda69e1
 - Commit a252bbf360532c39eb3d0264a7cbf3f80595b3a6
 - Commit 0c39c389db7100aab42192861507ea3b064e7edf
made by Brent.  The only commit not included is
Commit 40509e211cb0e7794c6b1b4f0f781446201a9a0b: since a Scanner is
meant to be used only by a single thread in a completely sequential
manner, and does not need any form of concurrent access protection,
I don't think making the region member "volatile" is necessary.